move flask implementation to root folder
